hengampilot/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
djangorestframework-simplejwt
drf_spectacular
django-cors-headers
celery
redis
coverage
Pillow


